requirements.txt
requests
feedparser
google-genai
resend
python-dotenv